requirements.txt
absl-py==0.9.0
appdirs==1.4.3
arrow==0.15.5
astroid==2.4.0
backcall==0.1.0
binaryornot==0.4.4
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
daemonize==2.5.0
decorator==4.4.2
distlib==0.3.0
filelock==3.0.12
future==0.18.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
grpcio==1.28.1
guildai==0.7.0rc9
idna==2.9
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
jinja2-time==0.2.0
joblib==0.14.1
jupyter-client==6.1.3
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
ml-workflow
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.2.0.34
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
poyo==0.5.0
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pylint==2.5.0
python-dateutil==2.8.1
python-slugify==4.0.0
pytorch-ignite==0.3.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scikit-optimize==0.5.2
scipy==1.4.1
six==1.14.0
tabview==1.4.4
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
text-unidecode==1.3
toml==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.9
virtualenv==20.0.18
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
